requirements.txt
Django==6.0.3
django-filter>=23.1
django-simple-history==3.11.0
Faker>=20.0.0
google-genai>=1.0.0
gunicorn>=22.0.0
hypothesis>=6.0.0
pdfplumber==0.11.9
protobuf==7.34.0
psycopg2-binary>=2.9.9
pypdf==6.8.0
PyPDF2==3.0.1
python-dotenv==1.2.2
python-magic>=0.4.27
reportlab>=4.0.0
hypothesis>=6.0.0
mkdocs>=1.6.0
mkdocs-material>=9.5.0
mkdocstrings[python]>=0.27.0
pymdown-extensions>=10.14
